pyproject.toml
[tool.poetry]
name = "iDemographer-sim"
version = "0.1.0"
description = ""
authors = ["Sharar, Abinav"]

[tool.poetry.dependencies]
python = "^3.6"
tensorflow = "^2.4.1"
pixellib = "^0.6.1"
selenium = "^3.141.0"
psycopg2-binary = "^2.8.6"
rich = "^9.13.0"
aspose-imaging-cloud = "^21.2.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
